+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## [8.6.0] - 2023-03-20
 ### Stable Release
 - Migrate Media to Common
+- Migrate LocaleMiddleware to Common
 
 ## [8.5.1] - 2023-03-12
 ### Stable Release